HSM-14-19 decomposition cont.: carve SketchDiagram + AppSettings + WorkbenchUI (app 7,674 -> 4,899)
Three more verbatim carves out of MeetingCaptureApp.swift into App/MeetingCapture/:
- SketchDiagram.swift (571) — the Pencil-as-a-diagram-language feature (HSM-14-08/09:
  geometry + on-device Vision + the LAN vision model -> native Mermaid).
- AppSettings.swift (210) — the Settings surface (where intelligence runs).
- WorkbenchUI.swift (1,994) — THE WORKBENCH, the app-within-the-app (the HSM-14-15 linear
  node-graph + the HSM-14-16 Blueprints canvas + inspectors); its pure engine stays in
  RuntimeCore/Workbench/.

One shared private helper (artifactTypeLabel) widened to internal so WorkbenchUI shares it —
exactly the cross-file dependency the verified-build discipline is for (caught + fixed).

Device-arch xcodebuild BUILD SUCCEEDED after each carve. MeetingCaptureApp.swift is now 4,899
(8,042 at the start of the decomposition — six modules carved, the monster nearly halved).
